🚚 refactor: widgets 레이어 신설, components를 widgets로 이관
[변경 내용]
- widgets 레이어 신설, components 레이어 제거(폴더 소멸)
  - header + sidebar -> widgets/navigation (header/, sidebar/ 하위폴더 보존)
  - footer -> widgets/footer
- headerNavPolicy(네비게이션 정책/설정) -> shared/config로 이동
- 소비자(routes, intro) import 경로 갱신
- eslint boundaries 설정에서 components element 및 관련 정책 제거,
  레이어 방향을 app > routes > widgets > features > entities > shared로 정리

[변경 이유]
- header/sidebar/footer는 도메인(entities/member)과 feature(auth)를 아는 조립 UI라
  범용 shared/ui가 아니라 widgets가 정확한 레이어
- header가 sidebar의 모바일 드로어(MobileSidebarDrawerContent)와 뷰 스위처를
  조합하므로 둘은 하나의 네비게이션 시스템. widgets/navigation 단일 슬라이스로
  합쳐 header->sidebar를 슬라이스 내부 참조로 전환(같은 slice="navigation")
- headerNavPolicy는 순수 네비게이션 정책이라 header 위젯과 intro 랜딩이 함께 사용.
  shared/config로 내려 intro(feature)->header(widget) 역방향 위반 제거
- boundaries 위반 30 -> 27건 (-3): components 대상 위반(header<->sidebar 2,
  intro->header 1)이 모두 해소. 남은 27건은 전부 features->features(수평)

[고려 사항]
- header/sidebar를 병합하되 하위폴더(header/, sidebar/)를 보존해 파일명 충돌
  방지 및 조직 유지. slice 캡처는 첫 하위폴더(navigation)라 둘 다 같은 슬라이스
- footer는 shared만 의존하는 독립 위젯이라 별도 슬라이스
- sidebar->header 역참조가 없어(단방향) 병합 안전
- components 레이어가 완전히 비워져 boundaries element에서 제거. 이로써 레이어가
  표준 FSD 5단(app/pages/widgets/features/entities/shared)으로 정규화
- 남은 features->features 27건(settings->auth 9, application->project UI 누수 7
  등)은 수평 결합으로 가장 낮은 심각도. 추가 entity 추출/ UI 재배치 대상
- ProjectApplyModal.ui.test 5건 실패는 기존 이슈로 무관
