+++ eslint.config.ts
@@ -78,7 +78,7 @@ export default tseslint.config(
   },
   {
     // FSD 레이어 경계 규칙. 의존 방향 (상위 -> 하위):
-    // app > routes(pages) > widgets|components > features > entities > shared|types
+    // app > routes(pages) > widgets > features > entities > shared|types
     // 하위 레이어는 상위 레이어를 import 할 수 없다. 베이스라인 확보를 위해 warn으로
     // 시작하고, 구조 개편 완료 후 error로 전환한다.
     files: ["src/**/*.{ts,tsx}"],
@@ -91,7 +91,6 @@ export default tseslint.config(
         { type: "app", pattern: "src/app" },
         { type: "routes", pattern: "src/routes" },
         { type: "widgets", pattern: "src/widgets/*", capture: ["slice"] },
-        { type: "components", pattern: "src/components/*", capture: ["slice"] },
         { type: "features", pattern: "src/features/*", capture: ["slice"] },
         { type: "entities", pattern: "src/entities/*", capture: ["slice"] },
         { type: "shared", pattern: "src/shared" },
@@ -117,7 +116,6 @@ export default tseslint.config(
               ],
               allow: [
                 { to: { element: { type: "widgets" } } },
-                { to: { element: { type: "components" } } },
                 { to: { element: { type: "features" } } },
                 { to: { element: { type: "entities" } } },
                 { to: { element: { type: "shared" } } },
@@ -125,23 +123,12 @@ export default tseslint.config(
               ],
             },
             {
-              from: [
-                { element: { type: "widgets" } },
-                { element: { type: "components" } },
-              ],
+              from: [{ element: { type: "widgets" } }],
               allow: [
                 {
                   to: {
                     element: {
                       type: "widgets",
-                      captured: { slice: "{{ from.element.captured.slice }}" },
-                    },
-                  },
-                },
-                {
-                  to: {
-                    element: {
-                      type: "components",
                       captured: { slice: "{{ from.element.captured.slice }}" },
                     },
                   },
