+++ src/widgets/footer/Footer.tsx
