+++ src/widgets/navigation/header/GenerationListItem.tsx
