+++ src/widgets/navigation/header/Profile.tsx
