+++ src/widgets/navigation/sidebar/MobileSidebarDrawerContent.tsx
