+++ src/widgets/navigation/sidebar/SideBarViewSwitcher.tsx
