+++ src/widgets/navigation/sidebar/menu/SideBarMenu.tsx
