+++ src/widgets/navigation/sidebar/menu/SideBarMenuItem.tsx
