+++ src/widgets/navigation/sidebar/menu/SideBarMenuTitle.tsx
