+++ src/widgets/navigation/sidebar/useVisibleSidebarSections.ts
