+++ src/widgets/navigation/sidebar/utils.test.ts
