+++ src/widgets/navigation/sidebar/utils.ts
